# Speedup `remove_redudant_moves()` by using a `HashSet` instead of a `BTreeSet`