@mohammadfawaz 2022-10-22
> 🚀 🚀 🚀
> 
> Slow down @mohammadfawaz, you're gonna beat my old DCA win!

Oh your fix was way more subtle than these low hanging fruits.